+++ packages/oc-template-es6-compiler/CHANGELOG.md
@@ -1,5 +1,13 @@
 # oc-template-es6-compiler
 
+## 6.0.1
+
+### Patch Changes
+
+- Allow users to augment ocevents
+- Updated dependencies
+  - oc-vite-compiler@4.2.9
+
 ## 6.0.0
 
 ### Major Changes