+++ packages/oc-template-esm-compiler/CHANGELOG.md
@@ -1,5 +1,13 @@
 # oc-template-esm-compiler
 
+## 0.2.1
+
+### Patch Changes
+
+- Allow users to augment ocevents
+- Updated dependencies
+  - oc-vite-compiler@4.2.9
+
 ## 0.2.0
 
 ### Minor Changes